Public Bug 173: Intermittent test failures due to race condition
In DiscoveryTest.EndpointListenerImpl the methods
endpointAdded(EndpointDescription, String) and
endpointRemoved(EndpointDescription, String) acquire and release the
synchronization semaphore regardless of the actual EndpointDescription
received. This leads to a race condition in testDiscovery122_6 causing
intermittent test failures and happens when an EndpointDescription
provider notifies EndpointListeners asynchronously (of previous
EndpointDescriptions with identical
EndpointListener.ENDPOINT_LISTENER_SCOPE).

EndpointListenerImpl should therefore check if the EndpointDescription
is indeed the one it is looking for and only then release the semaphore.

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>