Fix for assumption in test cases that the *first* supported config type
i.e. in the existing test code, supportedConfigTypes[0] is the config
type to use for the actual remote service export.  It is completely
possible that there will be more than one supportedConfigType available
in a given rs implementation, and that the first one not be the one
intended for use in the test.  The changes look for and expect to find
the loadCTProperties service.exported.configs property value and use
that config type as the one to export the test remote service.

Signed-off-by: scottslewis <slewis@composent.com>